renamed my Database package to AdapterModels
