+++ app/src/main/java/edu/uco/schambers/classmate/Adapter/AuthAdapter.java
@@ -1,19 +1,12 @@
 package edu.uco.schambers.classmate.Adapter;
 
 import android.app.Activity;
-import android.app.Fragment;
-import android.content.Context;
 import android.content.SharedPreferences;
 import android.util.Log;
 
 import org.json.JSONException;
-import org.json.JSONObject;
 
 import java.io.IOException;
-
-import edu.uco.schambers.classmate.Database.User;
-import edu.uco.schambers.classmate.Fragments.StudentInterface;
-import edu.uco.schambers.classmate.Fragments.TeacherInterface;
 
 /**
  * Created by calitova on 9/22/2015.
