+++ app/src/main/java/edu/uco/schambers/classmate/AdapterModels/DataRepo.java
@@ -1,4 +1,4 @@
-package edu.uco.schambers.classmate.Database;
+package edu.uco.schambers.classmate.AdapterModels;
 
 import android.content.ContentValues;
 import android.content.Context;
